requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
Django==2.2.7
django-annoying==0.10.6
django-crispy-forms==1.7.2
django-filter==2.2.0
django-jsonstore==0.4.1
django-material==1.6.0
django-material-admin==1.6.18
django-viewflow==1.6.0
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pytz==2019.3
redis==3.3.11
six==1.13.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
